Convertion functions moved to converters.py module
